tech(api): move EmptyAnswerError to shared
